requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
psycopg[binary]>=3.2.2,<3.3
PyYAML==6.0.2
